requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.21.0,<2.0.0
matplotlib>=3.4.0
Pillow>=8.0.0